Replace inline Stripe forms with SubscribeButton everywhere
Create SubscribeButton.svelte: on iOS native shows RevenueCat paywall,
on web submits Stripe form. Replace all 10 inline paywall forms/links
across stories, speak, sentences, review, and dialect section pages.
PaywallModal now uses SubscribeButton internally.
