+++ src/lib/components/PaywallModal.svelte
@@ -1,10 +1,7 @@
 <script lang="ts">
-  import { PUBLIC_PRICE_ID } from '$env/static/public'
   import Modal from '$lib/components/Modal.svelte';
   import Checkmark from '$lib/components/Checkmark.svelte';
-  import Button from '$lib/components/Button.svelte';
-  import { onMount } from 'svelte';
-  import { invalidateAll } from '$app/navigation';
+  import SubscribeButton from '$lib/components/SubscribeButton.svelte';
 
   type Props = {
     isOpen: boolean;
@@ -12,36 +9,6 @@
   }
 
   let { isOpen = false, handleCloseModal = () => {} }: Props = $props();
-
-  let isNative = $state(false);
-  let isPurchasing = $state(false);
-  let purchaseError = $state<string | null>(null);
-
-  onMount(() => {
-    isNative = !!(window as any).Capacitor?.isNativePlatform?.();
-  });
-
-  async function handleApplePurchase() {
-    isPurchasing = true;
-    purchaseError = null;
-    try {
-      const { RevenueCatUI } = await import('@revenuecat/purchases-capacitor-ui');
-      const result = await RevenueCatUI.presentPaywall();
-
-      if (result.paywallResult === 'PURCHASED' || result.paywallResult === 'RESTORED') {
-        // Immediately sync to DB
-        await fetch('/api/verify-apple-purchase', { method: 'POST' });
-        await invalidateAll();
-        handleCloseModal();
-      }
-    } catch (err: any) {
-      if (err?.code !== 'PURCHASE_CANCELLED') {
-        purchaseError = `[DEBUG] code=${err?.code} message=${err?.message} full=${JSON.stringify(err)}`;
-      }
-    } finally {
-      isPurchasing = false;
-    }
-  }
 </script>
 
 <Modal isOpen={isOpen} handleCloseModal={handleCloseModal} width="min(90%, 600px)" height="fit-content">
@@ -97,27 +64,7 @@
 
     <!-- Footer / Action -->
     <div class="p-6 pt-0">
-      {#if purchaseError}
-        <p class="text-red-400 text-sm mb-3 text-center">{purchaseError}</p>
-      {/if}
-
-      {#if isNative}
-        <Button
-          type="button"
-          onClick={handleApplePurchase}
-          disabled={isPurchasing}
-          className="!py-3 !text-lg w-full shadow-md"
-        >
-          {isPurchasing ? 'Loading...' : 'Subscribe Now'}
-        </Button>
-      {:else}
-        <form method="POST" action="/?/subscribe">
-          <input type="hidden" name="price_id" value={PUBLIC_PRICE_ID} />
-          <Button type="submit" className="!py-3 !text-lg w-full shadow-md">
-            Subscribe Now
-          </Button>
-        </form>
-      {/if}
+      <SubscribeButton className="!py-3 !text-lg w-full shadow-md" />
     </div>
   </div>
 </Modal>
